LPD-25362 Fix NPE. The listFiles() method may return null when the file is not a directory or when it's deleted at the same time by someone else. The null check existed earlier but wasn't kept when refactoring this method from using recursive to loop.

+++ portal-impl/src/com/liferay/portal/util/FileImpl.java
@@ -272,7 +272,13 @@ public class FileImpl implements com.liferay.portal.kernel.util.File {
 			File curDirectory;
 
 			while ((curDirectory = visitQueue.poll()) != null) {
-				for (File file : curDirectory.listFiles()) {
+				File[] files = curDirectory.listFiles();
+
+				if (files == null) {
+					continue;
+				}
+
+				for (File file : files) {
 					if (file.isFile()) {
 						file.delete();
 					}
